A few things
So torchick and decoy chickens are hexes that can be swapped in when the
command for that is made. Sot he function exists. Torchick requires a
decoy chicken is spawned there.
When a wall is spawned, the player sees the hex outline for half a
second, then the wall spawns as normal.